PWN-6255 - comment fixes
