+++ .idea/codeStyles/Project.xml
@@ -1,7 +1,6 @@
 <component name="ProjectCodeStyleConfiguration">
   <code_scheme name="Project" version="173">
     <option name="RIGHT_MARGIN" value="130" />
-    <option name="ENABLE_SECOND_REFORMAT" value="true" />
     <option name="SOFT_MARGINS" value="120" />
     <AndroidXmlCodeStyleSettings>
       <option name="LAYOUT_SETTINGS">
