+++ app/src/main/java/org/p2p/wallet/moonpay/repository/MoonpayBuyRemoteRepository.kt
@@ -1,19 +1,21 @@
 package org.p2p.wallet.moonpay.repository
 
 import org.p2p.core.token.Token
+import org.p2p.core.utils.Constants.SOL_SYMBOL
+import org.p2p.wallet.BuildConfig
 import org.p2p.wallet.infrastructure.network.data.ErrorCode
 import org.p2p.wallet.infrastructure.network.data.ServerException
 import org.p2p.wallet.moonpay.api.MoonpayApi
 import org.p2p.wallet.moonpay.api.MoonpayIpAddressResponse
 import org.p2p.wallet.moonpay.model.MoonpayBuyResult
-import org.p2p.core.utils.Constants.SOL_SYMBOL
 import java.math.BigDecimal
 
 class MoonpayBuyRemoteRepository(
     private val api: MoonpayApi,
-    private val moonpayApiKey: String,
     private val mapper: MoonpayApiMapper
 ) : MoonpayBuyRepository {
+
+    private val moonpayApiKey: String = BuildConfig.moonpayKey
 
     override suspend fun getBuyCurrencyData(
         baseCurrencyAmount: String?,
